editing changes continuation
