+++ src/main/java/db/DAO/ITourDAO.java
@@ -1,4 +1,4 @@
-package db.dao;
+package db.DAO;
 
 import db.models.Tour;
