+++ src/main/java/db/DAO/mysql/CityDAO.java
@@ -1,7 +1,7 @@
-package db.dao.mysql;
+package db.DAO.mysql;
 
 import db.ConnectionPool.ConnectionPool;
-import db.dao.ICityDAO;
+import db.DAO.ICityDAO;
 import db.models.City;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -12,24 +12,22 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-public class CityDao implements ICityDAO {
-    private static final Logger LOGGER = LogManager.getLogger(CityDao.class);
+public class CityDAO implements ICityDAO {
+    private static final Logger LOGGER = LogManager.getLogger(CityDAO.class);
+    private ConnectionPool connectionPool = ConnectionPool.getInstance();
 
     public List<City> getAllEntities(){
         List<City> cityList = new ArrayList<>();
         String sql = "SELECT * FROM cities";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try {
             PreparedStatement ps = con.prepareStatement(sql);
             ResultSet rs = ps.executeQuery();
-
             while (rs.next()) {
                 City city = new City();
                 city.setCityId(rs.getInt(1));
                 city.setCityName(rs.getString(2));
                 city.setCountryId(rs.getInt(3));
-
                 cityList.add(city);
             }
         } catch (Exception e) {
@@ -37,25 +35,22 @@ public class CityDao implements ICityDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return cityList;
     }
 
     public City getEntityById(int id) {
         City city = null;
-        Connection con = ConnectionPool.getInstance().getConnection();
+        Connection con = connectionPool.getConnection();
         String sql = "SELECT * FROM cities WHERE company_id = (?)";
-
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
-
             if (rs.next()) {
                 int cityId = rs.getInt("city_id");
                 String cityName = rs.getString("city_name");
@@ -68,51 +63,43 @@ public class CityDao implements ICityDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return city;
     }
 
     public City createEntity(City city) {
-        Connection con = ConnectionPool.getInstance().getConnection();
+        Connection con = connectionPool.getConnection();;
         String sql = "INSERT INTO cities (city_id, city_name, country_id) VALUES (?,?,?)";
-
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-
             ps.setInt(1, city.getCityId());
             ps.setString(2, city.getCityName());
             ps.setInt(3, city.getCountryId());
             ps.executeUpdate();
-
             LOGGER.info("Insertion was successful");
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return null;
     }
 
     public void updateEntity(City city) {
         String sql = "UPDATE cities SET city_name = (?), country_id = (?) WHERE city_id = (?)";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-
             LOGGER.info("in update customer: " + city);
-
             ps.setString(1, city.getCityName());
             ps.setInt(2, city.getCountryId());
             ps.setInt(3, city.getCityId());
@@ -122,7 +109,7 @@ public class CityDao implements ICityDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
@@ -132,20 +119,17 @@ public class CityDao implements ICityDAO {
 
     public void removeEntity(int id) {
         String sql = "Delete FROM cities WHERE city_id = (?)";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)){
-
             ps.setInt(1, id);
             ps.executeUpdate();
-
             LOGGER.info("Removal was successful");
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
