+++ src/main/java/db/DAO/mysql/CountryDAO.java
@@ -1,7 +1,7 @@
-package db.dao.mysql;
+package db.DAO.mysql;
 
 import db.ConnectionPool.ConnectionPool;
-import db.dao.ICountryDAO;
+import db.DAO.ICountryDAO;
 import db.models.Country;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -12,24 +12,21 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-public class CountryDao implements ICountryDAO {
-    private static final Logger LOGGER = LogManager.getLogger(CountryDao.class);
+public class CountryDAO implements ICountryDAO {
+    private static final Logger LOGGER = LogManager.getLogger(CountryDAO.class);
+    private ConnectionPool connectionPool = ConnectionPool.getInstance();
 
     public List<Country> getAllEntities(){
         List<Country> countryList = new ArrayList<>();
         String sql = "SELECT * FROM countries";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();;
         try {
-
             PreparedStatement ps = con.prepareStatement(sql);
             ResultSet rs = ps.executeQuery();
-
             while (rs.next()) {
                 Country country = new Country();
                 country.setCountryId(rs.getInt(1));
                 country.setCountryName(rs.getString(2));
-
                 countryList.add(country);
             }
         } catch (Exception e) {
@@ -37,29 +34,25 @@ public class CountryDao implements ICountryDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return countryList;
     }
 
     public Country getEntityById(int id) {
         Country country = null;
-        Connection con = ConnectionPool.getInstance().getConnection();
+        Connection con = connectionPool.getConnection();
         String sql = "SELECT * FROM countries WHERE country_id = (?)";
-
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
-
             if (rs.next()) {
                 int countryId = rs.getInt("country_id");
                 String countryName = rs.getString("country_name");
-
                 country = new Country(countryId, countryName);
             }
         } catch (Exception e) {
@@ -67,50 +60,42 @@ public class CountryDao implements ICountryDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return country;
     }
 
     public Country createEntity(Country country) {
-        Connection con = ConnectionPool.getInstance().getConnection();
+        Connection con = connectionPool.getConnection();
         String sql = "INSERT INTO countries (country_id, country_name) VALUES (?,?)";
-
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-
             ps.setInt(1, country.getCountryId());
             ps.setString(2, country.getCountryName());
             ps.executeUpdate();
-
             LOGGER.info("Insertion was successful");
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return null;
     }
 
     public void updateEntity(Country country) {
         String sql = "UPDATE countries SET country_name = (?) WHERE country_id = (?)";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-
             LOGGER.info("in update customer: " + country);
-
             ps.setString(1, country.getCountryName());
             ps.setInt(2, country.getCountryId());
             ps.execute();
@@ -119,7 +104,7 @@ public class CountryDao implements ICountryDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
@@ -129,20 +114,17 @@ public class CountryDao implements ICountryDAO {
 
     public void removeEntity(int id) {
         String sql = "Delete FROM countries WHERE country_id = (?)";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)){
-
             ps.setInt(1, id);
             ps.executeUpdate();
-
             LOGGER.info("Removal was successful");
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
