+++ src/main/java/db/DAO/mysql/CustomerDAO.java
@@ -1,7 +1,7 @@
-package db.dao.mysql;
+package db.DAO.mysql;
 
 import db.ConnectionPool.ConnectionPool;
-import db.dao.ICustomerDAO;
+import db.DAO.ICustomerDAO;
 import db.models.Customer;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -12,27 +12,24 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-public class CustomerDao implements ICustomerDAO {
+public class CustomerDAO implements ICustomerDAO {
 
-    private static final Logger LOGGER = LogManager.getLogger(CustomerDao.class);
+    private static final Logger LOGGER = LogManager.getLogger(CustomerDAO.class);
+    private ConnectionPool connectionPool = ConnectionPool.getInstance();
 
     public List<Customer> getAllEntities(){
         List<Customer> customerList = new ArrayList<>();
         String sql = "SELECT customer_id, first_name, last_name, phone FROM customers";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try {
-
             PreparedStatement ps = con.prepareStatement(sql);
             ResultSet rs = ps.executeQuery();
-
             while (rs.next()) {
                 Customer customer = new Customer();
                 customer.setCustomerId(rs.getInt(1));
                 customer.setFirstName(rs.getString(2));
                 customer.setLastName(rs.getString(3));
                 customer.setPhone(rs.getString(4));
-
                 customerList.add(customer);
             }
         } catch (Exception e) {
@@ -40,7 +37,7 @@ public class CustomerDao implements ICustomerDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
@@ -52,19 +49,16 @@ public class CustomerDao implements ICustomerDAO {
 
    public Customer getEntityById(int id) {
         Customer customers = null;
-        Connection con = ConnectionPool.getInstance().getConnection();
+        Connection con = connectionPool.getConnection();
         String sql = "SELECT * FROM customers WHERE customer_id = (?)";
-
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
-
             if (rs.next()) {
                 int customerId = rs.getInt("customer_id");
                 String firstName = rs.getString("first_name");
                 String lastName = rs.getString("last_name");
                 String phone = rs.getString("phone");
-
                 customers = new Customer(customerId, firstName, lastName, phone);
             }
         } catch (Exception e) {
@@ -72,53 +66,44 @@ public class CustomerDao implements ICustomerDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return customers;
     }
 
     public Customer createEntity(Customer customers) {
-        Connection con = ConnectionPool.getInstance().getConnection();
+        Connection con = connectionPool.getConnection();
         String sql = "INSERT INTO customers (customer_id, first_name, last_name, phone) VALUES (?,?,?,?)";
-
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-
             ps.setInt(1, customers.getCustomerId());
             ps.setString(2, customers.getFirstName());
             ps.setString(3, customers.getLastName());
             ps.setString(4, customers.getPhone());
-
             ps.executeUpdate();
-
             LOGGER.info("Insertion was successful");
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return null;
     }
 
     public void updateEntity(Customer customers) {
         String sql = "UPDATE customers SET first_name = (?), last_name = (?), phone = (?) WHERE customer_id = (?)";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-
             LOGGER.info("in update customer: " + customers);
-
             ps.setString(1, customers.getFirstName());
             ps.setString(2, customers.getLastName());
             ps.setString(3, customers.getPhone());
@@ -129,7 +114,7 @@ public class CustomerDao implements ICustomerDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
@@ -139,20 +124,17 @@ public class CustomerDao implements ICustomerDAO {
 
     public void removeEntity(int id) {
         String sql = "Delete FROM customers WHERE customer_id = ?";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)){
-
             ps.setInt(1, id);
             ps.executeUpdate();
-
             LOGGER.info("Removal was successful");
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
