+++ src/main/java/db/DAO/mysql/FlightDAO.java
@@ -1,7 +1,7 @@
-package db.dao.mysql;
+package db.DAO.mysql;
 
 import db.ConnectionPool.ConnectionPool;
-import db.dao.IFlightDAO;
+import db.DAO.IFlightDAO;
 import db.models.Flight;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -12,19 +12,17 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-public class FlightDao implements IFlightDAO {
-    private static final Logger LOGGER = LogManager.getLogger(FlightDao.class);
+public class FlightDAO implements IFlightDAO {
+    private static final Logger LOGGER = LogManager.getLogger(FlightDAO.class);
+    private ConnectionPool connectionPool = ConnectionPool.getInstance();
 
     public List<Flight> getAllEntities(){
         List<Flight>  flightList = new ArrayList<>();
         String sql = "SELECT * FROM flights";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try {
-
             PreparedStatement ps = con.prepareStatement(sql);
             ResultSet rs = ps.executeQuery();
-
             while (rs.next()) {
                 Flight flight = new Flight();
                 flight.setFlightId(rs.getInt(1));
@@ -32,7 +30,6 @@ public class FlightDao implements IFlightDAO {
                 flight.setAirlineId(rs.getInt(3));
                 flight.setOriginCityId(rs.getInt(4));
                 flight.setDestinationCityId(rs.getInt(5));
-
                 flightList.add(flight);
             }
         } catch (Exception e) {
@@ -40,32 +37,28 @@ public class FlightDao implements IFlightDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return flightList;
     }
 
     public Flight getEntityById(int id) {
         Flight flight = null;
-        Connection con = ConnectionPool.getInstance().getConnection();
+        Connection con = connectionPool.getConnection();
         String sql = "SELECT * FROM flights WHERE flight_id = (?)";
-
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
-
             if (rs.next()) {
                 int flightId = rs.getInt("flight_id");
                 double price = rs.getDouble("price");
                 int airlineId = rs.getInt("airline_id");
                 int originCityId = rs.getInt("origin_city_id");
                 int destinationCityId = rs.getInt("destination_city_id");
-
                 flight = new Flight(flightId, price, airlineId, originCityId, destinationCityId);
             }
         } catch (Exception e) {
@@ -73,54 +66,46 @@ public class FlightDao implements IFlightDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return flight;
     }
 
     public Flight createEntity(Flight flight) {
-        Connection con = ConnectionPool.getInstance().getConnection();
+        Connection con = connectionPool.getConnection();
         String sql = "INSERT INTO flights (flight_id, price, airline_id, origin_city_id, destination_city_id) VALUES (?,?,?,?,?)";
-
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-
             ps.setInt(1, flight.getFlightId());
             ps.setDouble(2, flight.getPrice());
             ps.setInt(3, flight.getAirlineId());
             ps.setInt(4, flight.getOriginCityId());
             ps.setInt(5, flight.getDestinationCityId());
             ps.executeUpdate();
-
             LOGGER.info("Insertion was successful");
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return null;
     }
 
     public void updateEntity(Flight flight) {
         String sql = "UPDATE flights SET price = (?), airline_id = (?), origin_city_id = (?), " +
                      "destination_city_id = (?) WHERE flight_id = (?)";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-
             LOGGER.info("in update customer: " + flight);
-
             ps.setDouble(1, flight.getPrice());
             ps.setInt(2, flight.getAirlineId());
             ps.setInt(3, flight.getOriginCityId());
@@ -132,7 +117,7 @@ public class FlightDao implements IFlightDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
@@ -142,20 +127,17 @@ public class FlightDao implements IFlightDAO {
 
     public void removeEntity(int id) {
         String sql = "Delete FROM flights WHERE flight_id = (?)";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)){
-
             ps.setInt(1, id);
             ps.executeUpdate();
-
             LOGGER.info("Removal was successful");
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
