+++ src/main/java/db/DAO/mysql/HotelDAO.java
@@ -1,8 +1,8 @@
-package db.dao.mysql;
+package db.DAO.mysql;
 
 import db.ConnectionPool.ConnectionPool;
-import db.dao.IDrivingCompanyDAO;
-import db.models.DrivingCompany;
+import db.DAO.IHotelDAO;
+import db.models.Hotel;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -12,120 +12,108 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-public class DrivingCompanyDao implements IDrivingCompanyDAO {
-    private static final Logger LOGGER = LogManager.getLogger(DrivingCompanyDao.class);
+public class HotelDAO implements IHotelDAO {
+    private static final Logger LOGGER = LogManager.getLogger(HotelDAO.class);
+    private ConnectionPool connectionPool = ConnectionPool.getInstance();
 
-    public List<DrivingCompany> getAllEntities(){
-
-        List<DrivingCompany>  dcList = new ArrayList<>();
-        String sql = "SELECT * FROM driving_companies";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+    public List<Hotel> getAllEntities(){
+        List<Hotel>  hotelList = new ArrayList<>();
+        String sql = "SELECT * FROM hotels";
+        Connection con = connectionPool.getConnection();
         try {
-
             PreparedStatement ps = con.prepareStatement(sql);
             ResultSet rs = ps.executeQuery();
-
             while (rs.next()) {
-                DrivingCompany dc = new DrivingCompany();
-                dc.setDrivingId(rs.getInt(1));
-                dc.setPrice(rs.getDouble(2));
-                dc.setCityId(rs.getInt(3));
-
-                dcList.add(dc);
+                Hotel hotel = new Hotel();
+                hotel.setHotelId(rs.getInt(1));
+                hotel.setHotelName(rs.getString(2));
+                hotel.setPrice(rs.getDouble(3));
+                hotel.setCityId(rs.getInt(4));
+                hotelList.add(hotel);
             }
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
-        return dcList;
+        return hotelList;
     }
 
-    public DrivingCompany getEntityById(int id) {
-
-        DrivingCompany dc = null;
-        Connection con = ConnectionPool.getInstance().getConnection();
-        String sql = "SELECT * FROM driving_companies WHERE driving_id = (?)";
-
+    public Hotel getEntityById(int id) {
+        Hotel hotel = null;
+        Connection con = connectionPool.getConnection();
+        String sql = "SELECT * FROM hotels WHERE hotel_id = (?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
-
             if (rs.next()) {
-                int drivingId = rs.getInt("driving_id");
+                int hotelId = rs.getInt("hotel_id");
+                String hotelName = rs.getString("hotel_name");
                 double price = rs.getDouble("price");
                 int cityId = rs.getInt("city_id");
-
-                dc = new DrivingCompany(drivingId, price, cityId);
+                hotel = new Hotel(hotelId, hotelName, price, cityId);
             }
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
-        return dc;
+        return hotel;
     }
 
-    public DrivingCompany createEntity(DrivingCompany dc) {
-        Connection con = ConnectionPool.getInstance().getConnection();
-        String sql = "INSERT INTO driving_companies (driving_id, price, city_id) VALUES (?,?,?)";
-
+    public Hotel createEntity(Hotel hotel) {
+        Connection con = connectionPool.getConnection();
+        String sql = "INSERT INTO hotels (hotel_id, hotel_name, price, city_id) VALUES (?,?,?,?)";
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-
-            ps.setInt(1, dc.getDrivingId());
-            ps.setDouble(2, dc.getPrice());
-            ps.setInt(3, dc.getCityId());
+            ps.setInt(1, hotel.getHotelId());
+            ps.setString(2, hotel.getHotelName());
+            ps.setDouble(3, hotel.getPrice());
+            ps.setInt(4, hotel.getCityId());
             ps.executeUpdate();
-
             LOGGER.info("Insertion was successful");
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return null;
     }
 
-    public void updateEntity(DrivingCompany dc) {
-        String sql = "UPDATE driving_companies SET price = (?), city_id = (?) WHERE driving_id = (?)";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+    public void updateEntity(Hotel hotel) {
+        String sql = "UPDATE hotels SET hotel_name = (?), price = (?), city_id = (?), " +
+                "WHERE hotel_id = (?)";
+        Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-
-            LOGGER.info("in update customer: " + dc);
-
-            ps.setDouble(1, dc.getPrice());
-            ps.setInt(2, dc.getCityId());
-            ps.setInt(3, dc.getDrivingId());
+            LOGGER.info("in update customer: " + hotel);
+            ps.setString(1, hotel.getHotelName());
+            ps.setDouble(2, hotel.getPrice());
+            ps.setInt(3, hotel.getCityId());
+            ps.setInt(4, hotel.getCityId());
             ps.execute();
         } catch(Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
@@ -134,21 +122,18 @@ public class DrivingCompanyDao implements IDrivingCompanyDAO {
     }
 
     public void removeEntity(int id) {
-        String sql = "Delete FROM driving_companies WHERE driving_id = (?)";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        String sql = "Delete FROM hotels WHERE hotel_id = (?)";
+        Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)){
-
             ps.setInt(1, id);
             ps.executeUpdate();
-
             LOGGER.info("Removal was successful");
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
