+++ src/main/java/db/DAO/mysql/OrderDAO.java
@@ -1,7 +1,7 @@
-package db.dao.mysql;
+package db.DAO.mysql;
 
 import db.ConnectionPool.ConnectionPool;
-import db.dao.IOrderDAO;
+import db.DAO.IOrderDAO;
 import db.models.Order;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -12,18 +12,17 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-public class OrderDao implements IOrderDAO {
-    private static final Logger LOGGER = LogManager.getLogger(OrderDao.class);
+public class OrderDAO implements IOrderDAO {
+    private static final Logger LOGGER = LogManager.getLogger(OrderDAO.class);
+    private ConnectionPool connectionPool = ConnectionPool.getInstance();
 
     public List<Order> getAllEntities(){
         List<Order>  orderList = new ArrayList<>();
         String sql = "SELECT * FROM orders";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try {
             PreparedStatement ps = con.prepareStatement(sql);
             ResultSet rs = ps.executeQuery();
-
             while (rs.next()) {
                 Order order = new Order();
                 order.setOrderId(rs.getInt(1));
@@ -34,7 +33,6 @@ public class OrderDao implements IOrderDAO {
                 order.setStatusId(rs.getInt(6));
                 order.setPaymentId(rs.getInt(7));
                 order.setDrivingCompaniesId(rs.getInt(8));
-
                 orderList.add(order);
             }
         } catch (Exception e) {
@@ -42,25 +40,22 @@ public class OrderDao implements IOrderDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return orderList;
     }
 
     public Order getEntityById(int id) {
         Order order = null;
-        Connection con = ConnectionPool.getInstance().getConnection();
+        Connection con = connectionPool.getConnection();
         String sql = "SELECT * FROM orders WHERE order_id = (?)";
-
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
-
             if (rs.next()) {
                 int orderId = rs.getInt("order_id");
                 double payment = rs.getDouble("payment");
@@ -70,7 +65,6 @@ public class OrderDao implements IOrderDAO {
                 int statusId = rs.getInt("status_id");
                 int paymentId = rs.getInt("payment_id");
                 int drivingId = rs.getInt("driving_id");
-
                 order = new Order(orderId, payment, customerId, staffId, toursId, statusId, paymentId, drivingId);
             }
         } catch (Exception e) {
@@ -78,23 +72,20 @@ public class OrderDao implements IOrderDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return order;
     }
 
     public Order createEntity(Order order) {
-        Connection con = ConnectionPool.getInstance().getConnection();
+        Connection con = connectionPool.getConnection();
         String sql = "INSERT INTO hotels (order_id, payment, customer_id, staff_id, tours_id, status_id, payment_id, driving_Id) " +
                      "VALUES (?,?,?,?,?,?,?,?)";
-
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-
             ps.setInt(1, order.getOrderId());
             ps.setDouble(2, order.getPayment());
             ps.setInt(3, order.getCustomerId());
@@ -104,32 +95,27 @@ public class OrderDao implements IOrderDAO {
             ps.setInt(7, order.getPaymentId());
             ps.setInt(8, order.getDrivingCompaniesId());
             ps.executeUpdate();
-
             LOGGER.info("Insertion was successful");
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return null;
     }
 
     public void updateEntity(Order order) {
         String sql = "UPDATE orders SET payment = (?), customer_id = (?), staff_id = (?), " +
                 "tours_id = ?, status_id = ?, payment_id = ?, driving_id = ? WHERE order_id = (?)";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
-
             LOGGER.info("in update customer: " + order);
-
             ps.setDouble(1, order.getPayment());
             ps.setInt(2, order.getCustomerId());
             ps.setInt(3, order.getStaffId());
@@ -144,7 +130,7 @@ public class OrderDao implements IOrderDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
@@ -154,20 +140,17 @@ public class OrderDao implements IOrderDAO {
 
     public void removeEntity(int id) {
         String sql = "Delete FROM orders WHERE order_id = (?)";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)){
-
             ps.setInt(1, id);
             ps.executeUpdate();
-
             LOGGER.info("Removal was successful");
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
