+++ src/main/java/db/DAO/mysql/TourDAO.java
@@ -1,7 +1,7 @@
-package db.dao.mysql;
+package db.DAO.mysql;
 
 import db.ConnectionPool.ConnectionPool;
-import db.dao.IToursDAO;
+import db.DAO.IToursDAO;
 import db.models.Tour;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -12,18 +12,17 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-public class TourDao implements IToursDAO {
-    private static final Logger LOGGER = LogManager.getLogger(TourDao.class);
+public class TourDAO implements IToursDAO {
+    private static final Logger LOGGER = LogManager.getLogger(TourDAO.class);
+    private ConnectionPool connectionPool = ConnectionPool.getInstance();
 
     public List<Tour> getAllEntities(){
         List<Tour>  tourList = new ArrayList<>();
         String sql = "SELECT * FROM tours";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try {
             PreparedStatement ps = con.prepareStatement(sql);
             ResultSet rs = ps.executeQuery();
-
             while (rs.next()) {
                 Tour tour = new Tour();
                 tour.setTourId(rs.getInt(1));
@@ -31,7 +30,6 @@ public class TourDao implements IToursDAO {
                 tour.setHotelId(rs.getInt(3));
                 tour.setFlightToId(rs.getInt(4));
                 tour.setFlightFromId(rs.getInt(5));
-
                 tourList.add(tour);
             }
         } catch (Exception e) {
@@ -39,32 +37,28 @@ public class TourDao implements IToursDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return tourList;
     }
 
     public Tour getEntityById(int id) {
         Tour tour = null;
-        Connection con = ConnectionPool.getInstance().getConnection();
+        Connection con = connectionPool.getConnection();
         String sql = "SELECT * FROM tours WHERE tour_id = (?)";
-
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
-
             if (rs.next()) {
                 int tourId = rs.getInt("tour_id");
                 String tourName = rs.getString("tour_name");
                 int hotelId = rs.getInt("hotel_id");
                 int flightToId = rs.getInt("flight_to_id");
                 int flightFromId = rs.getInt("flight_from_id");
-
                 tour = new Tour(tourId, tourName, hotelId,flightToId, flightFromId);
             }
         } catch (Exception e) {
@@ -72,21 +66,19 @@ public class TourDao implements IToursDAO {
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return tour;
     }
 
     public Tour createEntity(Tour tour) {
-        Connection con = ConnectionPool.getInstance().getConnection();
+        Connection con = connectionPool.getConnection();
         String sql = "INSERT INTO tours (tour_id, tour_name, hotel_id, flight_to_id, flight_from_id) " +
                 "VALUES (?,?,?,?,?)";
-
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             ps.setInt(1, tour.getTourId());
             ps.setString(2, tour.getTourName());
@@ -94,44 +86,39 @@ public class TourDao implements IToursDAO {
             ps.setInt(4, tour.getFlightToId());
             ps.setInt(5, tour.getFlightFromId());
             ps.executeUpdate();
-
             LOGGER.info("Insertion was successful");
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
             }
         }
-
         return null;
     }
 
     public void updateEntity(Tour tour) {
         String sql = "UPDATE tours SET tour_name = (?), hotel_id = (?)" +
                 "flight_to_id = (?), flight_from_id = (?) WHERE tour_id = (?)";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)) {
             LOGGER.info("in update customer: " + tour);
-
             ps.setString(1, tour.getTourName());
             ps.setInt(2, tour.getHotelId());
             ps.setInt(3, tour.getFlightToId());
             ps.setInt(4, tour.getFlightFromId());
             ps.setInt(5, tour.getTourId());
-
             ps.execute();
         } catch(Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
@@ -141,19 +128,17 @@ public class TourDao implements IToursDAO {
 
     public void removeEntity(int id) {
         String sql = "Delete FROM tours WHERE tour_id = (?)";
-        Connection con = ConnectionPool.getInstance().getConnection();
-
+        Connection con = connectionPool.getConnection();
         try (PreparedStatement ps = con.prepareStatement(sql)){
             ps.setInt(1, id);
             ps.executeUpdate();
-
             LOGGER.info("Removal was successful");
         } catch (Exception e) {
             LOGGER.error(e);
         } finally {
             if (con != null) {
                 try {
-                    ConnectionPool.getInstance().releaseConnection(con);
+                    connectionPool.releaseConnection(con);
                 } catch (Exception e) {
                     LOGGER.info(e);
                 }
